+++ actix-broker/src/broker.rs
@@ -50,22 +50,24 @@ impl Broker<ArbiterBroker> {
 
 /// The system service actor that keeps track of subscriptions and routes messages to them.
 impl<T> Broker<T> {
-    fn take_subs<M: BrokerMsg>(&mut self) -> Option<Vec<(TypeId, Recipient<M>)>> {
+    fn get_subs<M: BrokerMsg>(
+        &self,
+    ) -> Option<impl Iterator<Item = (usize, (&TypeId, &Recipient<M>))>> {
         let id = TypeId::of::<M>();
-        let subs = self.sub_map.get_mut(&id)?;
+        let msg_subs = self.sub_map.get(&id)?;
+
         trace!("Broker: Found subscription list for {:?}.", id);
-        let subs = subs
-            .drain(..)
-            .filter_map(|(id, s)| {
-                if let Ok(rec) = s.downcast::<Recipient<M>>() {
-                    Some((id, rec))
-                } else {
-                    None
-                }
-            })
-            .map(|(id, s)| (id, *s))
-            .collect();
-        Some(subs)
+
+        let iter = msg_subs
+            .iter()
+            .enumerate()
+            .map(|(idx, (actor_id, recipient))| {
+                let typed_recipient = recipient
+                    .downcast_ref::<Recipient<M>>()
+                    .expect("Message type should always be M");
+                (idx, (actor_id, typed_recipient))
+            });
+        Some(iter)
     }
 
     fn add_sub<M: BrokerMsg>(&mut self, sub: Recipient<M>, id: TypeId) {
@@ -79,6 +81,20 @@ impl<T> Broker<T> {
 
         trace!("Broker: Creating {:?} subscription list.", msg_id);
         self.sub_map.insert(msg_id, vec![(id, boxed)]);
+    }
+
+    fn remove_subs<M: BrokerMsg>(&mut self, indexes: &[usize]) {
+        if indexes.is_empty() {
+            return;
+        }
+
+        let msg_id = TypeId::of::<M>();
+        if let Some(subscribers) = self.sub_map.get_mut(&msg_id) {
+            trace!("Broker: Removing {:?} subscribers", indexes);
+            indexes.iter().rev().for_each(|&idx| {
+                subscribers.swap_remove(idx);
+            });
+        }
     }
 
     fn get_previous_msg<M: BrokerMsg>(&self) -> Option<M> {
@@ -127,24 +143,25 @@ impl<T: 'static + Unpin, M: BrokerMsg> Handler<IssueAsync<M>> for Broker<T> {
 
     fn handle(&mut self, msg: IssueAsync<M>, _ctx: &mut Context<Self>) {
         trace!("Broker: Received IssueAsync");
-        if let Some(mut subs) = self.take_subs::<M>() {
-            subs.drain(..).for_each(|(id, s)| {
-                if id == msg.1 {
-                    self.add_sub::<M>(s, id);
-                } else {
-                    match s.try_send(msg.0.clone()) {
-                        Ok(_) => self.add_sub::<M>(s, id),
-                        Err(SendError::Full(_)) => {
-                            // Ensure that that the message is delivered even if the mailbox is full.
-                            // We do a try first to remove receiver that have closed their mailbox.
-                            s.do_send(msg.0.clone());
-                            self.add_sub::<M>(s, id);
+        let subscriber_indexes = if let Some(subscribers) = self.get_subs::<M>() {
+            subscribers
+                .filter(|&(_, (actor_id, _))| !msg.1.eq(actor_id))
+                .filter_map(
+                    |(idx, (_, recipient))| match recipient.try_send(msg.0.clone()) {
+                        Err(SendError::Full(msg)) => {
+                            recipient.do_send(msg);
+                            None
                         }
-                        Err(_) => (),
-                    }
-                }
-            });
-        }
+                        Err(_) => Some(idx),
+                        _ => None,
+                    },
+                )
+                .collect()
+        } else {
+            vec![]
+        };
+
+        self.remove_subs::<M>(subscriber_indexes.as_slice());
         self.set_msg::<M>(msg.0);
     }
 }
@@ -154,18 +171,19 @@ impl<T: 'static + Unpin, M: BrokerMsg> Handler<IssueSync<M>> for Broker<T> {
 
     fn handle(&mut self, msg: IssueSync<M>, ctx: &mut Context<Self>) {
         trace!("Broker: Received IssueSync");
-        if let Some(mut subs) = self.take_subs::<M>() {
-            subs.drain(..).for_each(|(id, s)| {
-                if id == msg.1 {
-                    self.add_sub::<M>(s, id);
-                } else {
-                    s.send(msg.0.clone())
+
+        if let Some(subscribers) = self.get_subs::<M>() {
+            subscribers
+                .filter(|&(_, (actor_id, _))| !msg.1.eq(actor_id))
+                .for_each(|(_, (_, recipient))| {
+                    recipient
+                        .send(msg.0.clone())
                         .into_actor(self)
-                        .map(move |_, act, _| act.add_sub::<M>(s, id))
+                        .map(|_, _, _| ())
                         .wait(ctx);
-                }
-            });
+                });
         }
+
         self.set_msg::<M>(msg.0);
     }
 }