+++ actix-broker/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Improve broker message performance.
 - Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 0.4.3 - 2022-05-24